pyproject.toml
[tool.poetry]
name = "flow-with-duckdb"
version = "0.1.0"
description = ""
authors = ["kaushik911 <kaushikshamantha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
duckdb = "^1.0.0"
pyarrow = "^16.1.0"
pandas = "^2.2.2"
fire = "^0.6.0"
loguru = "^0.7.2"
pydantic = "^2.8.2"
pytest = "^8.2.2"
ruff = "^0.5.2"
google-cloud-bigquery = "^3.25.0"
google-auth = "^2.32.0"
google-cloud-bigquery-storage = "^2.25.0"
db-dtypes = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
